+++ README.md
@@ -49,26 +49,30 @@
   
 ## Building
 
-* All source code is located in the `src` directory
-  * Every function in the .py files within `src` contains docstrings so you can use help() to learn how they work
-  * `src/lambda_function.py` is the run-script used for deployment to AWS. It takes the place of `deployLOCAL/runETL.py` for running the code locally.
-    * Each run-script imports only the functions marked as `public API` in their docstrings. These functions are `Extract.extract_function`, `Transform.transform_function`, and `Load.load_function`. The helper functions within transform are not imported by the run-scripts.
+* All source code is located in the `myetl` directory
+  * Every function in the .py files within `myetl` contains docstrings so you can use help() to learn how they work
+  * `scripts/lambda_function.py` is the run-script used for deployment to AWS. It takes the place of `scripts/call_cli_handler.py` for running the code locally.
+    * Each run-script imports only the functions marked as `public API` in their docstrings. These functions are `extract.extract_function`, `transform.transform_function`, and `load.load_function`. The helper functions within transform are not imported by the run-scripts.
+  * Build the package by typing `pip install -r requirements.txt` from the root directory to run `setup.py` and install to your local machine.
 
 ## Running Locally
 
-* `deployLOCAL` contains an end-to-end run-script named `runETL.py`
-  * typing `python3 runETL.py -h` will display a help message that describes the arguments needed such as QUERY_DATE which is passed in as `-d YYYY-MM-DD`. I've also provided options to point to a local directory to store temporary files and delete it when the Lambda completes.
+* `scripts` contains an end-to-end run-script named `call_cli_handler.py`
+  * typing `python3 call_cli_handler.py -h` will display a help message that describes the arguments needed such as QUERY_DATE which is passed in as `-d YYYY-MM-DD`. I've also provided options to point to a local directory to store temporary files and delete it when the Lambda completes.
 
 ## Testing with pytest
 
-* From the root directory, type `pytest src`
-  * It is necessary to pass `src` because of all the code stored in `lib` that's needed for deployment to AWS. Amazon linux compatible pandas, numpy, and pytz have a few files with 'test' in their names and/or functions; pytest tries to test them and it goes rather poorly.
-  * Pytest will test the two helper functions in `src/Transform.py` using `src/test_transform.py`
+* From the root directory, type `pytest myetl`
+  * It is necessary to pass `myetl` because of all the code stored in `lib` that's needed for deployment to AWS. Amazon linux compatible pandas, numpy, and pytz have a few files with 'test' in their names and/or functions; pytest tries to test them and it goes rather poorly.
+  * Pytest will test the two helper functions in `myetl/transform.py` using `myetl/test_transform.py`
     * email to email_address is a direct mapping so no helper function was needed
   
 ## Deploying to AWS Lambda
 
-* In `deployAWS` there are two shell scripts
-  * `push2lambda.sh` zips up all necessary code and creates the Lambda using the AWS CLI
+* In `scripts` there are two shell scripts
+  * `bundle.sh` zips up all necessary code and creates the Lambda using the AWS CLI
   * `invoke_lambda.sh` will invoke the Lambda provided the user passes it one CL arg - date in YYYY-MM-DD as follows `invoke_lambda.sh YYYY-MM-DD`
   * Go check out `s3://mybucket1219` to view the output
+* UPDATE: code doesn't run in the Lambda anymore
+  * `bundle.sh` uses `pip install -r ` to install all dependencies of `myetl` which unfortunately installs a bunch of Mac specific packages; these break on the Lambda. I reverted back to using `scripts/push2lambda.sh` but because AWS Lambda doesn't recognize `myetl` as a package, all import statements of the form `from myetl import ` break.
+  * If you want to try my code in a Lambda, revert back to (any) prior commit in my repo and `push2lambda.sh` + `invoke_lambda.sh` will work.
