+++ requirements.txt
@@ -1,0 +1,6 @@
+botocore==1.12.130
+pandas==0.23.4
+numpy==1.15.4
+pytz==2018.7
+
+.
