+++ scripts/call_cli_handler.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Apr 15 10:06:48 2019
+
+@author: zackkingbackup
+"""
+
+from myetl import cli_handler
+
+cli_handler()
