+++ scripts/lambda_function.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Apr 10 16:27:42 2019
+
+@author: zackkingbackup
+"""
+
+from myetl import lambda_handler
