+++ scripts/outputfile.txt
@@ -1,1 +1,1 @@
-null
+{"errorMessage": "Unable to import module 'lambda_function': No module named 'myetl'", "errorType": "Runtime.ImportModuleError"}
